move still useful code out of old file and delete
